+++ worker/db.py
@@ -78,11 +78,11 @@ async def get_stats(db, config):
             ), decision_table AS (
                 SELECT 
                     review_id,
-                    CASE WHEN AVG(insult) > 2/3 THEN 1 ELSE 0 END insult,
-                    CASE WHEN AVG(severe_toxic) > 2/3 THEN 1 ELSE 0 END severe_toxic,
-                    CASE WHEN AVG(identity_hate) > 2/3 THEN 1 ELSE 0 END identity_hate,
-                    CASE WHEN AVG(threat) > 2/3 THEN 1 ELSE 0 END threat,
-                    CASE WHEN AVG(nsfw) > 2/3 THEN 1 ELSE 0 END nsfw
+                    CASE WHEN AVG(insult) > 2/3::float THEN 1 ELSE 0 END insult,
+                    CASE WHEN AVG(severe_toxic) > 2/3::float THEN 1 ELSE 0 END severe_toxic,
+                    CASE WHEN AVG(identity_hate) > 2/3::float THEN 1 ELSE 0 END identity_hate,
+                    CASE WHEN AVG(threat) > 2/3::float THEN 1 ELSE 0 END threat,
+                    CASE WHEN AVG(nsfw) > 2/3::float THEN 1 ELSE 0 END nsfw
                 FROM (
                         SELECT * FROM reviews_table
                         UNION
